refactor: extract shared searchWithFallback into linkupClient module
Both deepDiligence.ts and expandEntity.ts had near-identical
searchWithFallback() functions. Extracted into a shared
convex/domains/search/linkupClient.ts with SSRF validation,
bounded response reading, and Gemini grounding fallback.
deepDiligence gains SSRF + BOUND_READ protections it was missing.
